add prescription-id-driven async pipeline with typed upload/status DTOs and explicit PROCESSING/FAILED polling states

+++ src/main/java/com/TenaMed/verification/workflow/VerificationEngine.java
@@ -18,7 +18,7 @@ public class VerificationEngine {
 	}
 
 	@Value("${ocr.confidence-threshold}")
-	private int ocrConfidenceThreshold;
+	private double ocrConfidenceThreshold;
 
 	public boolean isLowConfidence(double confidence) {
 		return confidence < ocrConfidenceThreshold;
